Fix #21 Use non-zero ranges for diagnostics
If a FROM's base image's digest string is the empty string, the
generated diagnostic's range has a length of zero. This causes the
CLI to not highlight anything in its output. Use the entire
argument's range for the diagnostic instead so that the CLI can at
least highlight something meaningful.

Signed-off-by: Remy Suen <remy.suen@gmail.com>