Fix #2 Consider build stages when flagging duplicates
If a Dockerfile has multiple CMD, ENTRYPOINT, or HEALTHCHECK
instructions, a warning is currently created because only one of such
an instruction will be used by the Docker image. However, if each
instruction is contained within a different build stage, then no
errors should be flagged. This is because each stage is a distinct
Docker image so therea are not actually multiple instructions in one
image.

Signed-off-by: Remy Suen <remy.suen@gmail.com>