Fix #60 Do not validate variable subsitutions in RUN
The validator was primarily written to handle variable substitutions
in the Bash world but it seems like PowerShell does things quite
differently than Bash. Due to the wide range of differences that a
shell can choose to how to expand variables, the validator will simply
choose to skip RUN instructions over when validating variable
substitutions.

Signed-off-by: Remy Suen <remy.suen@gmail.com>