Synchronize supported Docker CE versions
The version in the CLI and the documentation does not match right now.
We should at least synchronize it up for now even if we are still way
off the mark. We will try to verify the latest version that we support
after updating the library with the recent changes in Docker CE.

Signed-off-by: Remy Suen <remy.suen@gmail.com>